requirements.txt
fastapi == 0.111.0
uvicorn == 0.30.1
langchain == 0.2.6
langchain-community == 0.2.6
python-dotenv == 1.0.1